fix gurkha/hammer split on us
