+++ config/splat.us.stnp3.yaml
@@ -104,8 +104,8 @@ segments:
       - [0x321EC, .rodata, giantbro_helpers] # Strings
       - [0x3224C, .rodata, 4E69C] # EntityHammer
       - [0x322CC, .rodata, 4E69C] # EntityGurkhaBodyParts
-      - [0x322F8, .rodata, 4F5B8] # EntityGurkha
-      - [0x32380, .rodata, 4F5B8] # EntityGurkhaSword
+      - [0x322F8, .rodata, 4F778] # EntityGurkha
+      - [0x32380, .rodata, 4F778] # EntityGurkhaSword
       - [0x323E4, .rodata, blade] # EntityBlade
       - [0x32464, .rodata, prim_helpers]
       - [0x3246C, c, e_room_bg]
@@ -150,7 +150,7 @@ segments:
       - [0x4D658, c, giantbro_helpers]
       - [0x4E04C, c]
       - [0x4E69C, c]
-      - [0x4F5B8, c]
+      - [0x4F778, c]
       - [0x50A00, c, blade]
       - [0x51F38, c, prim_helpers]
       - [0x52768, .bss, create_entity]
